Added functionality to get info to specificView
